64 bit frenzy

+++ header/generator.hpp
@@ -7,9 +7,9 @@
 #define POLYNOMIAL_GENERATOR
 namespace PolyGenerator
 {
-    Polynomial random(int);
+    Polynomial random(int64_t);
     Polynomial getCurrPoly();
-    std::vector<std::pair<int, int>> getCurrRoots();
+    std::vector<std::pair<int64_t, int64_t>> getCurrRoots();
 };
 
 #endif
