+++ header/grader.hpp
@@ -5,6 +5,6 @@
 #ifndef GRADING_METHOD
 #define GRADING_METHOD
 
-bool isAnswer(std::vector<std::pair<int, int>>, std::vector<std::string>);
+bool isAnswer(std::vector<std::pair<int64_t, int64_t>>, std::vector<std::string>);
 
 #endif
