+++ src/generator.cpp
@@ -9,16 +9,16 @@
 namespace PolyGenerator
 {
     Polynomial curr_poly;
-    std::vector<std::pair<int, int>> curr_roots;
+    std::vector<std::pair<int64_t, int64_t>> curr_roots;
 
-    Polynomial random(int degree)
+    Polynomial random(int64_t degree)
     {
-        std::vector<std::pair<int, int>> roots = randPoly(degree);
+        std::vector<std::pair<int64_t, int64_t>> roots = randPoly(degree);
 
-        int count = roots.size();
+        int64_t count = roots.size();
         std::vector<int64_t> base_vec({-roots[0].first, roots[0].second});
         Polynomial base(base_vec);
-        for (int i = 1; i < count; i++)
+        for (int64_t i = 1; i < count; i++)
         {
             std::vector<int64_t> temp_vec({-roots[i].first, roots[i].second});
             Polynomial temp(temp_vec);
@@ -34,7 +34,7 @@ namespace PolyGenerator
         return curr_poly;
     }
 
-    std::vector<std::pair<int, int>> getCurrRoots()
+    std::vector<std::pair<int64_t, int64_t>> getCurrRoots()
     {
         return curr_roots;
     }
